[ Fix ] SolutionList 데이터 캐싱 문제 (#398)
